+++ src/utils/__mocks__/npm-info.ts
@@ -1,5 +1,15 @@
 import { type PackageJson } from "../../types";
 
+export interface NpmInfoError {
+	code: string;
+	summary: string;
+	detail: string;
+}
+
+export function isNpmInfoError(error: unknown): error is NpmInfoError {
+	return Boolean(error && error instanceof Error && "summary" in error);
+}
+
 const mock = new Map<string, PackageJson>();
 let defaultInfo: PackageJson | null = null;
 
@@ -7,11 +17,11 @@ export function npmInfo(pkg: string): Promise<PackageJson> {
 	const mocked = mock.get(pkg);
 	if (mocked) {
 		return Promise.resolve(mocked);
-	} else if (defaultInfo) {
-		return Promise.resolve(defaultInfo);
-	} else {
-		throw new Error(`No mocked package data for ${pkg}`);
 	}
+	if (defaultInfo) {
+		return Promise.resolve(defaultInfo);
+	}
+	throw new Error(`No mocked package data for ${pkg}`);
 }
 
 export function npmInfoMockClear(): void {
@@ -23,5 +33,6 @@ export function npmInfoMockAdd(pkg: string, pkgData: PackageJson): void {
 }
 
 export function npmInfoMockDefault(pkgData: PackageJson): void {
+	/* eslint-disable-next-line unicorn/no-top-level-assignment-in-function -- technical debt, should use explicit state variable */
 	defaultInfo = { ...pkgData };
 }